10370: add test coverage for fix stale blocks new script
